+++ example/src/pages/PaginationDemo.vue
@@ -1,0 +1,16 @@
+<template>
+  <div>
+    <Card title="基础用法">
+      <Pagination />
+    </Card>
+
+    <Card title="其他用法">
+      <Pagination />
+    </Card>
+  </div>
+</template>
+
+<script setup>
+import { Pagination } from '@apathia/apathia'
+</script>
+