+++ packages/shared/src/index.ts
@@ -11,3 +11,5 @@ export * from './unrefElement'
 export * from './request'
 
 export * from './directives'
+
+export * from './merge'