+++ packages/alert/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @apathia/apathia.alert
 
+## 1.0.0
+
+### Major Changes
+
+- add pagination
+
+### Patch Changes
+
+- Updated dependencies
+  - @apathia/apathia.custom-render@1.0.0
+  - @apathia/apathia.icon@1.0.0
+  - @apathia/apathia.shared@1.0.0
+  - @apathia/apathia.twind@1.0.0
+
+## 0.0.11
+
+### Patch Changes
+
+- add table
+- Updated dependencies
+  - @apathia/apathia.custom-render@0.0.8
+  - @apathia/apathia.icon@0.0.8
+  - @apathia/apathia.shared@0.0.10
+  - @apathia/apathia.twind@0.0.8
+
 ## 0.0.10
 
 ### Patch Changes