+++ packages/confirm/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @apathia/apathia.confirm
 
+## 1.0.0
+
+### Major Changes
+
+- add pagination
+
+### Patch Changes
+
+- Updated dependencies
+  - @apathia/apathia.button@1.0.0
+  - @apathia/apathia.custom-render@1.0.0
+  - @apathia/apathia.modal@1.0.0
+  - @apathia/apathia.twind@1.0.0
+
+## 0.0.10
+
+### Patch Changes
+
+- add table
+- Updated dependencies
+  - @apathia/apathia.button@0.0.10
+  - @apathia/apathia.custom-render@0.0.8
+  - @apathia/apathia.modal@0.0.10
+  - @apathia/apathia.twind@0.0.8
+
 ## 0.0.9
 
 ### Patch Changes