+++ packages/custom-render/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @apathia/apathia.custom-render
 
+## 1.0.0
+
+### Major Changes
+
+- add pagination
+
+## 0.0.8
+
+### Patch Changes
+
+- add table
+
 ## 0.0.7
 
 ### Patch Changes