+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @apathia/apathia.shared
 
+## 1.0.0
+
+### Major Changes
+
+- add pagination
+
+### Patch Changes
+
+- Updated dependencies
+  - @apathia/apathia.twind@1.0.0
+
+## 0.0.10
+
+### Patch Changes
+
+- add table
+- Updated dependencies
+  - @apathia/apathia.twind@0.0.8
+
 ## 0.0.9
 
 ### Patch Changes