+++ packages/table/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @apathia1/apathia.table
 
+## 1.0.0
+
+### Major Changes
+
+- add pagination
+
+### Patch Changes
+
+- Updated dependencies
+  - @apathia/apathia.button@1.0.0
+  - @apathia/apathia.checkbox@1.0.0
+  - @apathia/apathia.custom-render@1.0.0
+  - @apathia/apathia.hooks@1.0.0
+  - @apathia/apathia.icon@1.0.0
+  - @apathia/apathia.loading@1.0.0
+  - @apathia/apathia.twind@1.0.0
+
+## 0.0.10
+
+### Patch Changes
+
+- add table
+- Updated dependencies
+  - @apathia/apathia.button@0.0.10
+  - @apathia/apathia.checkbox@0.0.10
+  - @apathia/apathia.custom-render@0.0.8
+  - @apathia/apathia.hooks@0.0.10
+  - @apathia/apathia.icon@0.0.8
+  - @apathia/apathia.loading@0.0.8
+  - @apathia/apathia.twind@0.0.8
+
 ## 0.0.9
 
 ### Patch Changes